mt5_api_server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
numpy==1.26.4
MetaTrader5==5.0.5328
watchdog==4.0.1
websockets==12.0
pandas==2.2.2
pyyaml==6.0.2
loguru==0.7.2
requests==2.32.3

